Derick's Evaluation test
a. set the server response to "terminating connection" whenever an Imap LOGOUT command is sent to the server from the client. e.g

 `Client: a LOGOUT
  Server: * terminating connection
  Server: a OK LOGOUT completed
 (Server and client then close the connection)`
b. Additionally, if the client remains idle, the Imap server should reply with "bye, initializing autologout."
c. For a NOOP command, the server should reply with "received a Noop command."
d. If the client sends a NOOP command, the server should execute the CHECK command, and if a CHECK command is sent, the server should reply with a NOOP command.